Replace all vxz prefixes with pgd.
The 'vxz' prefix was chosen at random to avoid any name collision with any
existing environment variables or functions. Now that we have a proper name for
the project, pgd seeems to be a reasonable name prefix, and it is very unlikely
to conflict with any other script/library's name.