Explicitly tell Git which .git/ directory to use.
If one was working on a Postrges plugin that's not part of the core, then he/she
was forced to change/set the $GIT_DIR variable so that the pgd functions could
determine where the Postgres sources were. Althoguh there were these helpful
functions pgdSetGitDir() and pgdUnsetGitDir() to do that, but it was still a
pain because the developer now lost the ability to work on their/plugin's own
Git-based sources.

Now we don't rely on the $GIT_DIR variable anymore, so although the developer
still has to use pgSetGitDir() to tell where Postgres sources are, he/she can
continue to work on their Git-based source code without having to call
pgdUnsetGitDir() anymore.